MovieSceneTracks: compute animation time using double precision in MapTimeToAnimation() for skeletal animation sections
This updates the MapTimeToAnimation() function for skeletal animation sections to compute
and return a double precision value for the animation time rather than single precision. This
helps to avoid floating point precision error when the returned value is used to construct an
FFrameTime.

Many clients that operate on time values for skeletal animation still work in single precision,
so most existing call sites were updated to cast the returned double back to a float.

#rb max.chen
#preflight 6192a36634377daf8610c225

#ROBOMERGE-OWNER: matt.johnson
#ROBOMERGE-AUTHOR: matt.johnson
#ROBOMERGE-SOURCE: CL 18208817 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v895-18170469)
#ROBOMERGE-CONFLICT from-shelf

[CL 18209363 by matt johnson in ue5-release-engine-staging branch]

+++ Engine/Source/Editor/MovieSceneTools/Private/TrackEditors/SkeletalAnimationTrackEditor.cpp
@@ -614,7 +614,7 @@ int32 FSkeletalAnimationSection::OnPaintSection( FSequencerSectionPainter& Paint
 		if (Section.GetRange().Contains(CurrentTime.FrameNumber) && Section.Params.Animation != nullptr)
 		{
 			// Draw the current time next to the scrub handle
-			const float AnimTime = Section.MapTimeToAnimation(CurrentTime, TickResolution);
+			const double AnimTime = Section.MapTimeToAnimation(CurrentTime, TickResolution);
 			const FFrameRate SamplingFrameRate = Section.Params.Animation->GetSamplingFrameRate();
 			const FFrameTime FrameTime = SamplingFrameRate.AsFrameTime(AnimTime);
